more robust destroy

+++ oslxc/commands/destroy.sh
@@ -11,11 +11,15 @@ if [ `whoami` != "root" ]; then
   exec sudo bash $CMD_PATH $@
 fi
 
-info "Removing the LVM volume"
-cexe "oslxc" "vgremove -f cinder-volumes"
+# If the container was not fully provisioned vgrename may not be there
+if cexe "oslxc" "which vgremove"; then
+  info "Removing the LVM volume"
+  # try to remove the volume group only if it's there
+  cexe "oslxc" "vgdisplay cinder-volumes" && \
+    cexe "oslxc" "vgremove -f cinder-volumes"
+  info "Cleanup the loop device"
+  (losetup -a | grep loop6) && losetup -d /dev/loop6 || true
+fi
 
 info "Destroying the container"
 lxc-destroy -n oslxc -f
-
-info "Cleanup the loop device"
-losetup -d /dev/loop6
